perf(game): Use std::to_chars for score-to-string conversion
The integer-to-string conversion for displaying player scores and
the tie count is updated from the {fmt} library's format_int to the
standard library's std::to_chars. This change improves performance
by eliminating heap allocations in the rendering loop and allows
std::to_chars to write directly into a stack-allocated buffer.

+++ source/player.cpp
@@ -60,7 +60,7 @@ u8 Player::GetSign() const
  * Get the player score.
  * @return Player score.
  */
-unsigned int Player::GetScore() const
+u16 Player::GetScore() const
 {
     return Score;
 }
